Move uv_available() to utils.
