+++ tests/e2e_tests/utils/e2e_utils.py
@@ -1,11 +1,25 @@
 """Utility functions for e2e tests."""
 
+import shlex
+import subprocess
+
+
+def uv_available():
+    """Ensure uv is available before running test."""
+    cmd = "uv self version -q"
+    cmd_parts = shlex.split(cmd)
+    try:
+        subprocess.run(cmd_parts)
+        return True
+    except FileNotFoundError:
+        # This is the exception raised on macOS when the command uv is unavailable.
+        return False
 
 
 def get_core_plugin_test_cmd(path_dsd, cli_options, platform_name_lower):
     """Get the command for running django-simple-deploy tests after a new plugin has been installed."""
     test_filename = f"test_{platform_name_lower}_config.py"
-    
+
     if cli_options.include_core_tests:
         # Run full set of core django-simple-deploy tests, and plugin integration tests.
         cmd = f"cd {path_dsd.as_posix()} && source .venv/bin/activate && pytest"
